Add a minimal (ObjC compatible) wrapper around EmrtdConnector
This is required to use the EmrtdConnector in cross-platform projects, such as with .net MAUI.

These cross platform solutions can't handle swift code at the moment, but only ObjC code.
This wrapper adds a minimal interface to use the EmrtdConnector from within ObjC, including using
completion handler for asynchronous operations instead of the typical delegate pattern
or the swift way with async/await.

The result is returned as JSON as String, which makes it more universally usable compared to "real objects".